backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
langchain
openai
scikit-learn
python-dotenv
